smoo-ublk: initial work (WIP)
Wiring up ublk requires interactions with the control plane (to create
devices) and individual device queues via io_uring.

The aim is to keep all the ublk + io_uring specifics confined to this
crate, and provide a high level interface for the I/O requests that the
main gadget driver cares about.

Signed-off-by: Sam Day <me@samcday.com>