gadget-cli: chest-burster the final big piece of the puzzle
We'd layered in the various necessary bits in previous commits. Now we
have a "real" smoo-gadget-cli that:

 * pokes usb-gadget to bring up configfs + functionfs
 * hands the epfile fds over to an instance of SmooGadget
 * wires up SmooUblk to pump it's i/o operations thru the gadget

Codex did this with minimal extra prompting. What it came up with
revealed an interesting limitation of the current
design/architecture/impl.

It decided that the gadget CLI needs a --blocks + --blocksize flags.
This is because we're not properly formalizing the host-initiated side
of the protocol (to "export" a block device from the host)

But this code already works more or less how I'm expecting with a quick
dummy_hcd test. So the next commits will firm it up into something just
about ready for 0.0.1

Signed-off-by: Sam Day <me@samcday.com>