treewide: Remove UBLK_F_USER_COPY semantics
At some point, Codex got confused about how we intend to manage the
buffers. I let it continue on so that we could scaffold out the main
shape of the code (thus serving as more context).

I prompted it back on the right path, more or less, in this commit.