dreams shattered. but we press on.
The DMA-BUF acceleration path doesn't seem like it's gonna work the
whole way. So we limit it to just the FunctionFS side now, and accept
that there's always a memcpy when passing the FFS <-> ublk boundary.

In future, we might be able to get ublk to work with this by allowing
DMA-BUFs to be pumped through its io loop.

Signed-off-by: Sam Day <me@samcday.com>