gadget-buffers: launched into the sun
The whole point of this abstraction was to give us a neat way to handle
the buffers we were hoping to share between ublk and ffs. But that's no
longer feasible (or at least, descoped from the 0.0.1 slice).

So we just fold the necessary bits of the DMA-BUF pokery into
gadget-core, and collapse the VecBufferPool bits into the ublk code.

Signed-off-by: Sam Day <me@samcday.com>