hop in nerds, we're going multi-export
finally got to the bottom of the wedging. all of the event handling and
reconciliation stuff was async-safe, but the thread processing ublk's
cmd ring was synchronous and blocking. So with multiple exports we end
up racing >1 ADD_DEV..START_DEV paths with each other, which results in
interleaved CQEs coming back, but we were only processing them one at a
time with blocking.

I'd say "stupid machines for writing that!"... But the tragic thing is,
that part of the ublk reactor is literally the only part I authored by
myself without Codex involvement. So. uh. Yeah. That sure is something.

Signed-off-by: Sam Day <me@samcday.com>