gadget: improved DMA-BUF usage
This brings us back to how we wanted (and briefly had) things to work.
If we are able to allocate and attach DMA-BUFs to the ffs epfiles, then
we *also* use those buffers in the ublk side.

I haven't *properly* verified this yet, but I think that this
means the data path is in the zero-copy state we intended.

Signed-off-by: Sam Day <me@samcday.com>