gadget: better liveness semantics
if we stop getting status polls from the host, we assume it's gone away
(or the link is dead).

when host is away, in-flight I/O is dumped and errors out with -EAGAIN.
as subsequent i/o comes in, we race it against a periodic liveness
check, and continue backing off consumers with -EAGAIN until host
returns.

Signed-off-by: Sam Day <me@samcday.com>